Add Fluent Validation

+++ CustomerProduct.Hosting.API/Controllers/CustomerController.cs
@@ -5,6 +5,8 @@ using CustomerProduct.Data.Entities;
 using CustomerProduct.Hosting.API.Common;
 using CustomerProduct.Hosting.API.Models.RequestModels;
 using CustomerProduct.Hosting.API.Models.ResponseModels;
+using CustomerProduct.Hosting.API.Validator;
+using FluentValidation.Results;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
@@ -33,29 +35,74 @@ namespace CustomerProduct.Hosting.API.Controllers
         [Route("selectproduct")]
         public ActionResult<ApiResponse<CustomerProductResponse>> Post([FromBody] CustomerProductCreateRequest request)
         {
-            var result = _customerProductRepository.InsertCustomerProduct(request.IdentificationNumber, request.ProductId);
+            ServicePrimitiveResponse insertCustomerProductResult = new ServicePrimitiveResponse();
 
-            if(result.ResponseCode == (int)EntityResponseCodes.Successfull)
+            //Fluent Validation Controls
+            CustomerProductValidator validator = new CustomerProductValidator();
+            ValidationResult results = validator.Validate(request);
+
+            if (results.IsValid)
             {
-                //return Ok(new ApiResponse { Code = (int)ApiResponseCodes.Ok });
-                return Get(request.IdentificationNumber);
-            }
+                //Check IdentificationNo IsExist
+                bool isExistCustomer = _customerRepository.FindFirstBy(p => p.IdentificationNumber.Equals(request.IdentificationNumber)).ResponseCode == (int)EntityResponseCodes.NoRecordFound;
+                if (isExistCustomer)
+                    return BadRequest(new ApiResponse { Code = (int)ApiResponseCodes.ValidationError, ErrorMessage = " Identification Number Is Not Exist!" });
 
-            return Ok(new ApiResponse { Code = (int)ApiResponseCodes.Ok });
+                //Check ProductId IsExist
+                bool isExistProduct = _productRepository.Find(request.ProductId).ResponseCode == (int)EntityResponseCodes.NoRecordFound;
+                if (isExistProduct)
+                    return BadRequest(new ApiResponse { Code = (int)ApiResponseCodes.ValidationError, ErrorMessage = "Product Id Is Not Exist!" });
+
+                insertCustomerProductResult = _customerProductRepository.InsertCustomerProduct(request.IdentificationNumber, request.ProductId);
+
+                if (insertCustomerProductResult.ResponseCode == (int)EntityResponseCodes.Successfull)
+                    return Get(request.IdentificationNumber); //If Success Call Get ProductInfo endpoint
+            }
+            else
+                return BadRequest(GetValidationErrors(results));
+
+            throw new ApiException((int)ApiResponseCodes.DbError, insertCustomerProductResult.ResponseMessage);
 
         }
 
-        [HttpGet]
-        [Route("{identificationNumber}/selectedproducts")]
-        public ActionResult<ApiResponse<CustomerProductResponse>> Get(string identificationNumber)
+
+        public ApiResponse GetValidationErrors(ValidationResult results)
         {
-            //Get Products For Customer Identification Number
+            //Handle Validation Errors
+            List<string> validationErrorList = new List<string>();
 
-            var serviceEntityResponse = _customerProductRepository.GetSelectedProductsByIdentificationNumber(identificationNumber);
-
-            if(serviceEntityResponse.ResponseCode == (int)Enums.EntityResponseCodes.Successfull)
+            foreach (var value in results.Errors)
             {
-                return Ok(new ApiResponse<IEnumerable<CustomerProductResponse>>((int)ApiResponseCodes.Ok, serviceEntityResponse.EntityDataList.Select(x => new CustomerProductResponse
+                validationErrorList.Add(value.ErrorMessage);
+            }
+
+            var response = new ApiResponse
+            {
+                Code = (int)ApiResponseCodes.ValidationError,
+                ValidationErrors = validationErrorList
+            };
+
+            return response;
+        }
+
+        [HttpGet]
+        [Route("selectedproducts")]
+        public ActionResult<ApiResponse<CustomerProductResponse>> Get([FromQuery] string identificationNumber)
+        {
+            //Check identificationNumber Field
+            if (string.IsNullOrEmpty(identificationNumber))
+                return BadRequest(new ApiResponse { Code = (int)ApiResponseCodes.ValidationError, ErrorMessage = "IdentificationNumber Required" });
+
+            //Check IdentificationNo IsExist
+            bool isExistCustomer = _customerRepository.FindFirstBy(p => p.IdentificationNumber.Equals(identificationNumber)).ResponseCode == (int)EntityResponseCodes.NoRecordFound;
+            if (isExistCustomer)
+                return BadRequest(new ApiResponse { Code = (int)ApiResponseCodes.ValidationError, ErrorMessage = " Identification Number Is Not Exist!" });
+
+            var serviceEntityProductResponse  = _customerProductRepository.GetSelectedProductsByIdentificationNumber(identificationNumber);
+
+            if(serviceEntityProductResponse.ResponseCode == (int)Enums.EntityResponseCodes.Successfull)
+            {
+                return Ok(new ApiResponse<IEnumerable<CustomerProductResponse>>((int)ApiResponseCodes.Ok, serviceEntityProductResponse.EntityDataList.Select(x => new CustomerProductResponse
                 {
                     ProductId = x.Product.ProductId,
                     ProductName = x.Product.ProductName,
@@ -64,7 +111,7 @@ namespace CustomerProduct.Hosting.API.Controllers
                 })));
             }
 
-            return NotFound();
+            throw new ApiException((int)ApiResponseCodes.DbError, serviceEntityProductResponse.ResponseMessage);
         }
 
 
@@ -72,17 +119,43 @@ namespace CustomerProduct.Hosting.API.Controllers
         [Route("unselectproduct")]
         public ActionResult<ApiResponse<int>> Delete([FromBody] CustomerProductDeleteRequest request)
         {
-            ServiceEntityResponse<Data.Entities.CustomerProduct> serviceEntityResponse = _customerProductRepository.FindFirstBy(p => p.IdentificationNumber.Equals(request.IdentificationNumber) && p.ProductId == request.ProductId);
+            ServicePrimitiveResponse deleteCustomerProductResponse = new ServicePrimitiveResponse();
 
-            if(serviceEntityResponse.ResponseCode == (int)Enums.EntityResponseCodes.Successfull)
+            //Fluent Validation Controls
+            CustomerProductValidator validator = new CustomerProductValidator();
+            ValidationResult results = validator.Validate(request);
+
+            if (results.IsValid)
             {
-                _customerProductRepository.Delete(serviceEntityResponse.EntityData);
-                _customerProductRepository.Save();
+                //Check IdentificationNo IsExist
+                bool isExistCustomer = _customerRepository.FindFirstBy(p => p.IdentificationNumber.Equals(request.IdentificationNumber)).ResponseCode == (int)EntityResponseCodes.NoRecordFound;
+                if (isExistCustomer)
+                    return BadRequest(new ApiResponse { Code = (int)ApiResponseCodes.ValidationError, ErrorMessage = " Identification Number Is Not Exist!" });
 
-                return Ok(new ApiResponse { Code = (int)ApiResponseCodes.Ok });
+                //Check ProductId IsExist
+                bool isExistProduct = _productRepository.Find(request.ProductId).ResponseCode == (int)EntityResponseCodes.NoRecordFound;
+                if (isExistProduct)
+                    return BadRequest(new ApiResponse { Code = (int)ApiResponseCodes.ValidationError, ErrorMessage = "Product Id Is Not Exist!" });
+
+
+                //Get CustomerProductInfo
+                ServiceEntityResponse<Data.Entities.CustomerProduct> serviceEntityResponse = _customerProductRepository.FindFirstBy(p => p.IdentificationNumber.Equals(request.IdentificationNumber) && p.ProductId == request.ProductId);
+
+                if (serviceEntityResponse.ResponseCode == (int)Enums.EntityResponseCodes.Successfull)
+                {
+                    _customerProductRepository.Delete(serviceEntityResponse.EntityData);
+                    
+                    deleteCustomerProductResponse = _customerProductRepository.Save();
+                    
+                    if(deleteCustomerProductResponse.ResponseCode == (int)EntityResponseCodes.Successfull)
+                        return Ok(new ApiResponse { Code = (int)ApiResponseCodes.Ok });
+                    else
+                        throw new ApiException((int)ApiResponseCodes.DbError, deleteCustomerProductResponse.ResponseMessage);
+                }
             }
+               
+            return BadRequest(GetValidationErrors(results));
 
-            return Ok(new ApiResponse<int>((int)ApiResponseCodes.Ok, Convert.ToInt32(1)));
         }
